Exercises 5.11 and 5.12

+++ src/App.js
@@ -1,16 +1,16 @@
-import React, {useEffect, useRef, useState} from 'react'
-import Blog from './components/Blog'
-import blogService from './services/blogs'
-import LoginForm from "./components/LoginForm";
-import UserDetail from "./components/UserDetail";
-import Notification from "./components/Notification";
-import Togglable from "./components/Togglable";
-import CreatBlogForm from "./components/CreateBlogForm";
+import React, {useEffect, useRef, useState} from "react"
+import Blog from "./components/Blog"
+import blogService from "./services/blogs"
+import LoginForm from "./components/LoginForm"
+import UserDetail from "./components/UserDetail"
+import Notification from "./components/Notification"
+import Togglable from "./components/Togglable"
+import CreatBlogForm from "./components/CreateBlogForm"
 
 const App = () => {
     const [blogs, setBlogs] = useState([])
-    const [username, setUsername] = useState('')
-    const [password, setPassword] = useState('')
+    const [username, setUsername] = useState("")
+    const [password, setPassword] = useState("")
     const [user, setUser] = useState(null)
     const [errorMessage, setErrorMessage] = useState(null)
     const [messageClass, setMessageClass] = useState(".error")
@@ -27,7 +27,7 @@ const App = () => {
     }, [])
 
     useEffect(() => {
-        const loggedUserJSON = window.localStorage.getItem('loggedBlogAppUser')
+        const loggedUserJSON = window.localStorage.getItem("loggedBlogAppUser")
         if (loggedUserJSON) {
             const user = JSON.parse(loggedUserJSON)
             setUser(user)
@@ -37,21 +37,24 @@ const App = () => {
 
     const compare = (a, b) => {
         if (a.likes < b.likes) {
-            return 1;
+            return 1
         }
         if (a.likes > b.likes) {
-            return -1;
+            return -1
         }
-        return 0;
+        return 0
     }
 
 
     const removeBlog = async (blog) => {
-        await blogService.remove(blog.id)
-        const newBlogs = blogs.filter((value) => {
-            return value.id !== blog.id
-        })
-        setBlogs(newBlogs)
+        const result = window.confirm(`Remove blog ${blog.title} by ${blog.author}`)
+        if (result) {
+            await blogService.remove(blog.id)
+            const newBlogs = blogs.filter((value) => {
+                return value.id !== blog.id
+            })
+            setBlogs(newBlogs)
+        }
     }
 
 
@@ -68,7 +71,7 @@ const App = () => {
 
             {user !== null && <Togglable buttonLabel="Create New Blog" ref={blogFormRef}>
                 <CreatBlogForm blogFormRef={blogFormRef} setBlogs={setBlogs} blogs={blogs}
-                               setErrorMessage={setErrorMessage} setMessageClass={setMessageClass}/>
+                    setErrorMessage={setErrorMessage} setMessageClass={setMessageClass}/>
             </Togglable>}
